+++ index.js
@@ -3,7 +3,4 @@ import { AppRegistry } from "react-native";
 import App from "./App";
 import { name as appName } from "./app.json";
 
-if (__DEV__) {
-  import("./ReactotronConfig").then(() => console.log("Reactotron Configured"));
-}
 AppRegistry.registerComponent(appName, () => App);